core/pyproject.toml
[build-system]
requires = [
    "setuptools==69.5.1",
    "setuptools_scm==8.1.0",
    "setuptools_dynamic_dependencies @ git+https://github.com/beeware/setuptools_dynamic_dependencies",
]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "toga-core"
description = "A Python native, OS native GUI toolkit."
readme = "README.rst"
requires-python = ">= 3.8"
license.text = "New BSD"
authors = [
    {name="Russell Keith-Magee", email="russell@keith-magee.com"},
]
maintainers = [
    {name="BeeWare Team", email="team@beeware.org"},
]
keywords = [
    "gui",
    "widget",
    "cross-platform",
    "toga",
    "desktop",
    "mobile",
    "web",
    "macOS",
    "cocoa",
    "iOS",
    "android",
    "windows",
    "winforms",
    "linux",
    "freeBSD",
    "gtk",
    "console",
    "web",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Software Development :: Widget Sets",
]
dependencies = [
    "travertino >= 0.3.0",
    # limited to <=3.9 for the `group` argument for `entry_points()`
    "importlib_metadata >= 4.4.0; python_version <= '3.9'",
]

[project.optional-dependencies]
# Extras used by developers *of* Toga are pinned to specific versions to
# ensure environment consistency.
dev = [
    "coverage[toml] == 7.5.1",
    "Pillow == 10.3.0",
    # Pre-commit 3.6.0 deprecated support for Python 3.8
    "pre-commit == 3.5.0 ; python_version < '3.9'",
    "pre-commit == 3.7.1 ; python_version >= '3.9'",
    "pytest == 8.2.0",
    "pytest-asyncio == 0.23.6",
    "pytest-freezer == 0.4.8",
    "setuptools-scm == 8.0.4",
    "tox == 4.15.0",
    # typing-extensions needed for TypeAlias added in Py 3.10
    "typing-extensions == 4.9.0 ; python_version < '3.10'",
]
docs = [
    "furo == 2024.5.6",
    "Pillow == 10.3.0",
    "pyenchant == 3.2.2",
    # Sphinx 7.2 deprecated support for Python 3.8
    "sphinx == 7.1.2 ; python_version < '3.9'",
    "sphinx == 7.3.7 ; python_version >= '3.9'",
    "sphinx_tabs == 3.4.5",
    # Sphinx 2024.2.4 deprecated support for Python 3.8
    "sphinx-autobuild == 2021.3.14 ; python_version < '3.9'",
    "sphinx-autobuild == 2024.4.16 ; python_version >= '3.9'",
    "sphinx-autodoc-typehints == 2.1.0",
    "sphinx-csv-filter == 0.4.1",
    "sphinx-copybutton == 0.5.2",
    "sphinx-toolbox == 3.5.0",
    "sphinxcontrib-spelling == 8.0.0",
]

[project.urls]
Homepage = "https://beeware.org/project/projects/libraries/toga/"
Funding = "https://beeware.org/contributing/membership/"
Documentation = "https://toga.readthedocs.io/"
Tracker = "https://github.com/beeware/toga/issues"
Source = "https://github.com/beeware/toga"
Changelog = "https://toga.readthedocs.io/en/stable/background/project/releases.html"

[project.entry-points."toga.image_formats"]
pil = "toga.plugins.image_formats.PILConverter"

[tool.setuptools_scm]
root = ".."

[tool.coverage.run]
parallel = true
branch = true
relative_files = true

# See notes in the root pyproject.toml file.
source = ["src"]
source_pkgs = ["toga"]

[tool.coverage.paths]
source = [
    "src/toga",
    "**/toga",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "error",
]
